+++ tests/MailCatcherTest.php
@@ -10,12 +10,26 @@ use Faker\Factory;
 /**
  * Test the MailCatcher
  *
- * @author Tobias van Beek <t.vanbeek@connexx.nl>
+ * @author Tobias van Beek <t.vanbeek@tjvb.nl>
  *
  * @group mailcatcher
  */
 class MailCatcherTest extends TestCase
 {
+    /**
+     *
+     * {@inheritDoc}
+     * @see \Orchestra\Testbench\TestCase::setUp()
+     */
+    protected function setUp()
+    {
+        parent::setUp();
+        // we need it for almost every test
+        \config(['mailcatchall.enabled' => true]);
+        // we don't need this for the most of the test so we disable it by default
+        \config(['mailcatchall.add_receivers_to_html' => false]);
+        \config(['mailcatchall.add_receivers_to_text' => false]);
+    }
 
     /**
      * Test that it will do nothing if catchmail isn't enabled
@@ -52,7 +66,6 @@ class MailCatcherTest extends TestCase
     {
         $originalConfig = \config('mailcatchall.enabled');
         $originalReceiver = \config('mailcatchall.receiver');
-        \config(['mailcatchall.enabled' => true]);
         \config(['mailcatchall.receiver' => null]);
 
         $catcher = new MailCatcher();
@@ -85,7 +98,6 @@ class MailCatcherTest extends TestCase
         $originalConfig = \config('mailcatchall.enabled');
         $originalReceiver = \config('mailcatchall.receiver');
         $receiver = $faker->email;
-        \config(['mailcatchall.enabled' => true]);
         \config(['mailcatchall.receiver' => $receiver]);
 
         $catcher = new MailCatcher();
@@ -96,8 +108,10 @@ class MailCatcherTest extends TestCase
             'setBcc',
         ]);
         $messageMock->shouldReceive('setTo')->once()->with($receiver);
+        $messageMock->shouldReceive('getTo');
         $messageMock->shouldReceive('getCc');
         $messageMock->shouldReceive('getBcc');
+        $messageMock->shouldReceive('getContentType');
 
         $eventMock->message = $messageMock;
 
@@ -118,7 +132,6 @@ class MailCatcherTest extends TestCase
         $originalConfig = \config('mailcatchall.enabled');
         $originalReceiver = \config('mailcatchall.receiver');
         $receiver = $faker->email;
-        \config(['mailcatchall.enabled' => true]);
         \config(['mailcatchall.receiver' => $receiver]);
 
         $catcher = new MailCatcher();
@@ -127,10 +140,12 @@ class MailCatcherTest extends TestCase
         $messageMock->shouldNotReceive([
             'setBcc',
         ]);
+        $messageMock->shouldReceive('getTo');
         $messageMock->shouldReceive('setTo')->once()->with($receiver);
         $messageMock->shouldReceive('setCc')->once()->with([]);
         $messageMock->shouldReceive('getCc')->andReturn([$faker->email]);
         $messageMock->shouldReceive('getBcc');
+        $messageMock->shouldReceive('getContentType');
 
         $eventMock->message = $messageMock;
 
@@ -151,7 +166,6 @@ class MailCatcherTest extends TestCase
         $originalConfig = \config('mailcatchall.enabled');
         $originalReceiver = \config('mailcatchall.receiver');
         $receiver = $faker->email;
-        \config(['mailcatchall.enabled' => true]);
         \config(['mailcatchall.receiver' => $receiver]);
 
         $catcher = new MailCatcher();
@@ -160,14 +174,239 @@ class MailCatcherTest extends TestCase
         $messageMock->shouldNotReceive([
             'setCc',
         ]);
+        $messageMock->shouldReceive('getTo');
         $messageMock->shouldReceive('setTo')->once()->with($receiver);
         $messageMock->shouldReceive('getCc');
         $messageMock->shouldReceive('getBcc')->andReturn([$faker->email]);
         $messageMock->shouldReceive('setBcc')->once()->with([]);
+        $messageMock->shouldReceive('getContentType');
 
         $eventMock->message = $messageMock;
 
         $catcher->catchmail($eventMock);
+
+        \config(['mailcatchall.enabled' => $originalConfig]);
+        \config(['mailcatchall.receiver' => $originalReceiver]);
+    }
+
+    /**
+     * Test that the to is set in the text view
+     *
+     * @tests
+     */
+    public function it_will_add_original_to_in_text_view()
+    {
+        $faker = Factory::create();
+        $originalConfig = \config('mailcatchall.enabled');
+        $originalReceiver = \config('mailcatchall.receiver');
+        $receiver = $faker->email;
+        $originalTo = $faker->email;
+        \config(['mailcatchall.receiver' => $receiver]);
+        \config(['mailcatchall.add_receivers_to_text' => true]);
+
+        $catcher = new MailCatcher();
+        $eventMock = \Mockery::mock(MessageSending::class);
+
+        $message = new \Swift_Message();
+        $message->setBody($faker->text, 'text');
+        $message->setTo($originalTo);
+
+        $eventMock->message = $message;
+
+        $catcher->catchmail($eventMock);
+
+        $this->assertContains($originalTo, $message->getBody());
+
+        \config(['mailcatchall.enabled' => $originalConfig]);
+        \config(['mailcatchall.receiver' => $originalReceiver]);
+    }
+
+    /**
+     * Test that the to is set in the html view
+     *
+     * @tests
+     */
+    public function it_will_add_original_to_in_html_view()
+    {
+        $faker = Factory::create();
+        $originalConfig = \config('mailcatchall.enabled');
+        $originalReceiver = \config('mailcatchall.receiver');
+        $receiver = $faker->email;
+        $originalTo = $faker->email;
+        \config(['mailcatchall.receiver' => $receiver]);
+        \config(['mailcatchall.add_receivers_to_html' => true]);
+
+        $catcher = new MailCatcher();
+        $eventMock = \Mockery::mock(MessageSending::class);
+
+        $message = new \Swift_Message();
+        $message->setBody($faker->text, 'html');
+        $message->setTo($originalTo);
+
+        $eventMock->message = $message;
+
+        $catcher->catchmail($eventMock);
+
+        $this->assertContains($originalTo, $message->getBody());
+
+        \config(['mailcatchall.enabled' => $originalConfig]);
+        \config(['mailcatchall.receiver' => $originalReceiver]);
+    }
+
+    /**
+     * Test that the to is not set in the html view if disabled
+     *
+     * @tests
+     */
+    public function it_will_not_add_original_to_in_html_view_if_disabled()
+    {
+        $faker = Factory::create();
+        $originalConfig = \config('mailcatchall.enabled');
+        $originalReceiver = \config('mailcatchall.receiver');
+        $receiver = $faker->email;
+        $originalTo = $faker->email;
+        \config(['mailcatchall.receiver' => $receiver]);
+
+        $catcher = new MailCatcher();
+        $eventMock = \Mockery::mock(MessageSending::class);
+
+        $message = new \Swift_Message();
+        $message->setBody($faker->text, 'html');
+        $message->setTo($originalTo);
+
+        $eventMock->message = $message;
+
+        $catcher->catchmail($eventMock);
+
+        $this->assertNotContains($originalTo, $message->getBody());
+
+        \config(['mailcatchall.enabled' => $originalConfig]);
+        \config(['mailcatchall.receiver' => $originalReceiver]);
+    }
+
+    /**
+     * Test that the cc is set in the text view
+     *
+     * @tests
+     */
+    public function it_will_add_original_cc_in_text_view()
+    {
+        $faker = Factory::create();
+        $originalConfig = \config('mailcatchall.enabled');
+        $originalReceiver = \config('mailcatchall.receiver');
+        $receiver = $faker->email;
+        $originalCC = $faker->email;
+        \config(['mailcatchall.receiver' => $receiver]);
+        \config(['mailcatchall.add_receivers_to_text' => true]);
+
+        $catcher = new MailCatcher();
+        $eventMock = \Mockery::mock(MessageSending::class);
+
+        $message = new \Swift_Message();
+        $message->setBody($faker->text, 'text');
+        $message->setCc($originalCC);
+
+        $eventMock->message = $message;
+
+        $catcher->catchmail($eventMock);
+
+        $this->assertContains($originalCC, $message->getBody());
+
+        \config(['mailcatchall.enabled' => $originalConfig]);
+        \config(['mailcatchall.receiver' => $originalReceiver]);
+    }
+
+    /**
+     * Test that the cc is set in the html view
+     *
+     * @tests
+     */
+    public function it_will_add_original_cc_in_html_view()
+    {
+        $faker = Factory::create();
+        $originalConfig = \config('mailcatchall.enabled');
+        $originalReceiver = \config('mailcatchall.receiver');
+        $receiver = $faker->email;
+        $originalCC = $faker->email;
+        \config(['mailcatchall.receiver' => $receiver]);
+        \config(['mailcatchall.add_receivers_to_html' => true]);
+
+        $catcher = new MailCatcher();
+        $eventMock = \Mockery::mock(MessageSending::class);
+
+        $message = new \Swift_Message();
+        $message->setBody($faker->text, 'html');
+        $message->setCc($originalCC);
+
+        $eventMock->message = $message;
+
+        $catcher->catchmail($eventMock);
+
+        $this->assertContains($originalCC, $message->getBody());
+
+        \config(['mailcatchall.enabled' => $originalConfig]);
+        \config(['mailcatchall.receiver' => $originalReceiver]);
+    }
+
+    /**
+     * Test that the bcc is set in the text view
+     *
+     * @tests
+     */
+    public function it_will_add_original_bcc_in_text_view()
+    {
+        $faker = Factory::create();
+        $originalConfig = \config('mailcatchall.enabled');
+        $originalReceiver = \config('mailcatchall.receiver');
+        $receiver = $faker->email;
+        $originalBcc = $faker->email;
+        \config(['mailcatchall.receiver' => $receiver]);
+        \config(['mailcatchall.add_receivers_to_text' => true]);
+
+        $catcher = new MailCatcher();
+        $eventMock = \Mockery::mock(MessageSending::class);
+
+        $message = new \Swift_Message();
+        $message->setBody($faker->text, 'text');
+        $message->setBcc($originalBcc);
+
+        $eventMock->message = $message;
+
+        $catcher->catchmail($eventMock);
+
+        $this->assertContains($originalBcc, $message->getBody());
+
+        \config(['mailcatchall.enabled' => $originalConfig]);
+        \config(['mailcatchall.receiver' => $originalReceiver]);
+    }
+
+    /**
+     * Test that the bcc is set in the html view
+     *
+     * @tests
+     */
+    public function it_will_add_original_bcc_in_html_view()
+    {
+        $faker = Factory::create();
+        $originalConfig = \config('mailcatchall.enabled');
+        $originalReceiver = \config('mailcatchall.receiver');
+        $receiver = $faker->email;
+        $originalBcc= $faker->email;
+        \config(['mailcatchall.receiver' => $receiver]);
+        \config(['mailcatchall.add_receivers_to_html' => true]);
+
+        $catcher = new MailCatcher();
+        $eventMock = \Mockery::mock(MessageSending::class);
+
+        $message = new \Swift_Message();
+        $message->setBody($faker->text, 'html');
+        $message->setBcc($originalBcc);
+
+        $eventMock->message = $message;
+
+        $catcher->catchmail($eventMock);
+
+        $this->assertContains($originalBcc, $message->getBody());
 
         \config(['mailcatchall.enabled' => $originalConfig]);
         \config(['mailcatchall.receiver' => $originalReceiver]);