+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+- Add the option to add the original to, cc and bcc to the email body
+
 
 ## 0.1.0
 - Add an event handling to the Laravel sended mail